+++ lino/core/__init__.py
@@ -1,4 +1,6 @@
 """
+The ``lino.core`` package
+=========================
 
 .. autosummary::
    :toctree: